Improve code for readability.
Essentially, this mitigates the document-global code
and wraps it into the "main()" function which is now
invoked when the document is ready. This makes the
execution flow easier to follow for newcomers to
the code base.

Also, dedicated functions "navigatePreviousHour()"
and "navigateNextHour()" have been introduced to
improve the API for navigating in time.

No other features have been added.