Navigate in time using keyboard.
This improves usability when exploring the data through
the time domain. 

The user is now able to navigate forward and backwards 
in time by using the arrow left/right keys.

While the default is to navigate by a single amount of
the selected resolution (changeUnit), holding down the 
shift and alt key or both expands that delta.

- Shift:		One day.
- Alt:			One week.
- Shift+Alt:	One month.